Rename, new specification links, removed case sensitivity
