+++ app/src/main/java/zapsolutions/zap/lnurl/pay/LnUrlPaySecondResponse.java
@@ -8,7 +8,7 @@ import zapsolutions.zap.lnurl.LnUrlResponse;
  * This class helps to work with the received response from a LNURL-pay request.
  * <p>
  * Please refer to step 6 in the following reference:
- * https://github.com/btcontract/lnurl-rfc/blob/master/lnurl-pay.md
+ * https://github.com/fiatjaf/lnurl-rfc/blob/luds/06.md
  */
 public class LnUrlPaySecondResponse extends LnUrlResponse implements Serializable {
